requirements.txt
Django==1.7.3
django-modeltranslation==0.8
pytz==2014.10
